move code indexer related code to a new package
