+++ modules/indexer/code/bleve_test.go
@@ -1,0 +1,5 @@
+// Copyright 2019 The Gitea Authors. All rights reserved.
+// Use of this source code is governed by a MIT-style
+// license that can be found in the LICENSE file.
+
+package code
